+++ tools/dfuse/__init__.py
@@ -1,0 +1,4 @@
+from dfuse.DfuDevice import DfuDevice
+from dfuse.DfuStatus import DfuStatus
+from dfuse.DfuState import DfuState
+from dfuse.DfuFile import DfuFile